+++ chess/interfaz.py
@@ -1,97 +1,116 @@
-from chess.piezas import Torre, Caballo, Alfil, Dama, Rey, Peon, BLANCO, NEGRO
-from chess.tablero import Tablero
-from chess.diccionarios import diccionarioSimbolos as simbolos
+from piezas import Torre, Caballo, Alfil, Dama, Rey, Peon, BLANCO, NEGRO
+from tablero import Tablero
+from diccionarios import diccionarioSimbolos as simbolos
+from diccionarios import diccionarioFilas as posiciones
 
 BLANCO='Blanco'
 NEGRO='Negro'
-CLEAR="\033[H\033[J"  # para limpiar la pantalla sin importar la libreria os
 class Juego:
     def __init__(self,mostrar_tablero=True):
+        #self.__tablero__=Tablero()
         self.__tablero__=Tablero()
         self.__turno_actual__=BLANCO
         self.__turno_color_inicial__=BLANCO
         self.__turno_color_segundo__=NEGRO
         self.__juego_finalizado__=False
         self.__piezas__=[]
-        self.__mostrar_tablero__=mostrar_tablero
         self.__contador_jugadas__=0
         self.__primera_jugada__=True
-        self.PonerPiezas()
-        self.__ganador__=""
-        #self.__tablero__.mostrar_tablero()
-        if self.__mostrar_tablero__: #para que cuando ejecute los tests del juego no se muestre el tablero
-            self.__tablero__.mostrar_tablero()
-    def PonerPiezas(self): #inicializa las piezas en sus posiciones y alamacena las piezas en una lista para poder determinar el color ganador mas adelante
+        self.__ganador__=None
+        
+      
+    def inicializarPiezas(self): #inicializa las piezas en sus posiciones y alamacena las piezas en una lista para poder determinar el color ganador mas adelante
         #Blancas
         
         self.__piezas__.extend([
-            Peon(self.__turno_color_inicial__, "a2",self.__tablero__), Peon(self.__turno_color_inicial__, "b2",self.__tablero__),
-            Peon(self.__turno_color_inicial__, "c2",self.__tablero__), Peon(self.__turno_color_inicial__, "d2",self.__tablero__),
-            Peon(self.__turno_color_inicial__, "e2",self.__tablero__), Peon(self.__turno_color_inicial__, "f2",self.__tablero__),
-            Peon(self.__turno_color_inicial__, "g2",self.__tablero__), Peon(self.__turno_color_inicial__, "h2",self.__tablero__),
-            Torre(self.__turno_color_inicial__, "a1",self.__tablero__), Torre(self.__turno_color_inicial__, "h1",self.__tablero__),
-            Caballo(self.__turno_color_inicial__, "b1",self.__tablero__), Caballo(self.__turno_color_inicial__, "g1",self.__tablero__),
-            Alfil(self.__turno_color_inicial__, "c1",self.__tablero__), Alfil(self.__turno_color_inicial__, "f1",self.__tablero__),            
-            Dama(self.__turno_color_inicial__, "e1",self.__tablero__), Rey(self.__turno_color_inicial__, "d1",self.__tablero__)
-        ])
+            # Peon(self.__turno_color_inicial__, (0,1)), Peon(self.__turno_color_inicial__, (1,1)),
+            # Peon(self.__turno_color_inicial__, (2,1)), Peon(self.__turno_color_inicial__, (3,1)),
+            # Peon(self.__turno_color_inicial__, (4,1)), Peon(self.__turno_color_inicial__, (5,1)),
+            # Peon(self.__turno_color_inicial__, (6,1)), Peon(self.__turno_color_inicial__, (7,1)),
+            Torre(self.__turno_color_inicial__, (0,0)), Torre(self.__turno_color_inicial__, (7,0)),
+            Caballo(self.__turno_color_inicial__, (1,0)), Caballo(self.__turno_color_inicial__, (6,0)),
+            Alfil(self.__turno_color_inicial__, (2,0)), Alfil(self.__turno_color_inicial__, (5,0)),            
+            Dama(self.__turno_color_inicial__, (3,0)), Rey(self.__turno_color_inicial__, (4,0))
+       ])
 
-        # Negras
         self.__piezas__.extend([
-            Peon(self.__turno_color_segundo__, "a7",self.__tablero__), Peon(self.__turno_color_segundo__, "b7",self.__tablero__),
-            Peon(self.__turno_color_segundo__, "c7",self.__tablero__), Peon(self.__turno_color_segundo__, "d7",self.__tablero__),
-            Peon(self.__turno_color_segundo__, "e7",self.__tablero__), Peon(self.__turno_color_segundo__, "f7",self.__tablero__),
-            Peon(self.__turno_color_segundo__, "g7",self.__tablero__), Peon(self.__turno_color_segundo__, "h7",self.__tablero__),
-            Torre(self.__turno_color_segundo__, "a8",self.__tablero__), Torre(self.__turno_color_segundo__, "h8",self.__tablero__),
-            Caballo(self.__turno_color_segundo__, "b8",self.__tablero__), Caballo(self.__turno_color_segundo__, "g8",self.__tablero__),
-            Alfil(self.__turno_color_segundo__, "c8",self.__tablero__), Alfil(self.__turno_color_segundo__, "f8",self.__tablero__),            
-            Dama(self.__turno_color_segundo__, "e8",self.__tablero__), Rey(self.__turno_color_segundo__, "d8",self.__tablero__)
+            Peon(self.__turno_color_segundo__, (0,6)), Peon(self.__turno_color_segundo__, (1,6)),
+            # Peon(self.__turno_color_segundo__, (2,6)), Peon(self.__turno_color_segundo__, (3,6)),
+            # Peon(self.__turno_color_segundo__, (4,6)), Peon(self.__turno_color_segundo__, (5,6)), 
+            # Peon(self.__turno_color_segundo__, (6,6)), Peon(self.__turno_color_segundo__, (7,6)),
+            # Torre(self.__turno_color_segundo__, (0,7)), Torre(self.__turno_color_segundo__, (7,7)),
+            # Caballo(self.__turno_color_segundo__, (1,7)), Caballo(self.__turno_color_segundo__, (6,7)),
+            # Alfil(self.__turno_color_segundo__, (2,7)), Alfil(self.__turno_color_segundo__, (5,7)),            
+            # Dama(self.__turno_color_segundo__, (3,7)), Rey(self.__turno_color_segundo__, (4,7))
             
         ])
-    def obtenerCoordenadas(self):
-        coordenadas = None
-        while coordenadas is None:
-            try:
-                input_pos_pieza = input("Movimiento: ")
-                coordenadas = input_pos_pieza.split(" ")
-                if len(coordenadas) != 2:
-                    raise ValueError("Tenes que ingresar exactamente 2 coordenadas.")
-
-                for coord in coordenadas:
-                    if len(coord) != 2 or coord[0] not in "abcdefgh" or coord[1] not in "12345678":
-                        raise ValueError("Coordenada invalida.")
-                        
-            except ValueError as e:
-                print(e)
-                coordenadas = None 
-
-        return coordenadas[0], coordenadas[1]
-
-    def procesarJuego(self):
-        if self.__juego_finalizado__:
-            return 
-        posicion_inicial, posicion_final = self.obtenerCoordenadas()
-        if self.moverPieza(posicion_inicial, posicion_final):
-         
-            return True
-        return False
-    def mostrarGanador(self):
-        if self.determinarGanador():
-            print(f"El jugador {self.__ganador__} ha ganado")
-            self.__juego_finalizado__ = True
-    def moverPieza(self, posicion_inicial, posicion_final): #recibe posicion inicial y final, almacena en pieza la pieza que retorna determinarPieza
-        pieza=self.determinarPieza(posicion_inicial)        #Y en base a la posicion inicial que le pasemos selecciona la pieza y la mueve a la posicion final y aumenta el contador para el cambio de turno
-        if pieza:
-            posicion_original = pieza.posicion
-            if pieza.mover(posicion_final,self.__tablero__):
-                self.__contador_jugadas__+=1
-                return True
-            pieza.establecerPosicion(posicion_original)
-        return False
-    def determinarPieza(self, posicion): #se le pasa como parametro de la posicion de la pieza que queremos conocer y retorna la pieza en el caso de que la posicion sea correcta
+    def agregarPiezasEnTablero(self):
         for pieza in self.__piezas__:
-            if pieza.posicion==posicion and pieza.color==self.__turno_actual__:
+            self.__tablero__.insertarEnTableroEnInicializacion(pieza)
+   
+    def determinarPieza(self, posicion): #se le pasa como parametro de la posicion de la pieza que queremos conocer y retorna la pieza en el caso de que la posicion sea correcta
+        columna,fila=posicion
+        for pieza in self.__piezas__:
+            if (pieza.columna==columna and pieza.fila==fila) and pieza.color==self.__turno_actual__:
                 return pieza
         return False
+    def determinarPiezaDestino(self,posicion):
+        columna,fila=posicion
+        for pieza in self.__piezas__:
+            if (pieza.columna==columna and pieza.fila==fila) and pieza.color!=self.__turno_actual__:
+                return pieza
+        return False
+    def moverPieza(self, posicion_inicial, nueva_posicion):
+        pieza = self.determinarPieza(posicion_inicial)
+        if not pieza:
+            return False
+        columna_inicial, fila_inicial = posicion_inicial
+        columna_final, fila_final = nueva_posicion
+        pieza_destino=self.determinarPiezaDestino(nueva_posicion)
+        camino = pieza.checkMovimiento(nueva_posicion)
+        if camino and self.__tablero__.checkCamino(pieza, camino) and self.__tablero__.agregarEnTablero(pieza, nueva_posicion):
+            # Mover la pieza en el tablero
+            self.__tablero__.agregarEnTablero(pieza, nueva_posicion)
+            if pieza_destino:
+                self.__piezas__.remove(pieza_destino)
+
+            self.__contador_jugadas__+=1
+            if isinstance(pieza, Peon):
+                pieza.__movido__=True
+            return True
+        else:
+            pieza.establecerPosicion((columna_inicial, fila_inicial))
+            return False
+    def ProcesarInput(self):
+        coordenadas = None
+        while coordenadas is None:
+            coordenadas = input("Ingresar movimiento: ").replace(" ","").strip()
+            if coordenadas=="exit":
+                self.__juego_finalizado__=True
+                return False
+                
+
+            if self.excepcion(coordenadas):
+                primerLetra = posiciones[coordenadas[0].lower()]
+                primerNumero = int(coordenadas[1]) - 1
+                segundaLetra = posiciones[coordenadas[2].lower()]
+                segundoNumero = int(coordenadas[3]) - 1
+                return (primerLetra, primerNumero), (segundaLetra, segundoNumero)
+            coordenadas = None
+
+    def excepcion(self, coordenadas):
+        try:
+            coord = coordenadas.replace(" ", "").lower()
+            if len(coord) != 4:
+                raise ValueError("Se esperan 2 coordenadas(columna y fila)")
+            letras, numeros = 'abcdefgh', '12345678'
+            if (coord[0] not in letras or coord[2] not in letras or coord[1] not in numeros or coord[3] not in numeros):
+                raise ValueError("Formato incorrecto")
+            return True
+        except ValueError as e:
+            print(f"Error: {e}")
+            return False
+
     def determinarGanador(self):#itera sobre todas las posiciones del tablero, y si no hay piezas de un color, se da como ganador el otro color
         piezas_blanco = False
         piezas_negro = False
@@ -104,27 +123,20 @@ class Juego:
                 elif pieza in simbolos[NEGRO].values():
                     piezas_negro = True
                 
-                if piezas_blanco and piezas_negro:
-                    return None #
-        
         if not piezas_negro:
-            #print("El jugador blanco ha ganado")
-            self.__ganador__=BLANCO
-            return True
+            return BLANCO   
         elif not piezas_blanco:
-            #print("El jugador negro ha ganado")
-            self.__ganador__=NEGRO
-            return True
-        return False
+            return NEGRO
+        return None
     def turnos(self): #Si contador_jugadas es par el turno es de blancas, sino de negras y llama a procesarJuego
-        if self.determinarGanador():
-            print(f"El jugador {self.__ganador__} ha ganado")
+        if self.determinarGanador() is not None:
+            self.__ganador__=self.determinarGanador()
+            #print(f"El jugador {self.__ganador__} ha ganado")
             self.__juego_finalizado__ = True
             
         if self.__contador_jugadas__%2==0:
             self.__turno_actual__=BLANCO
         elif self.__contador_jugadas__%2==1:
             self.__turno_actual__=NEGRO
-        print(f"Turno de {self.__turno_actual__}")  
-       
+        print(f"Turno de {self.__turno_actual__}")