+++ CHANGELOG.md
@@ -1,4 +1,25 @@
 # CHANGELOG
+## [2.0.2] - 2024-08-27
+### Arreglar
+- Archivo `chess/piezas.py`
+- Archivo `chess/interfaz.py`
+## [2.0.1] - 2024-08-26
+### Actualizar
+- Archivo `chess/main.py`
+- Archivo `chess/interfaz.py`
+## [2.0.0] - 2024-08-24
+### Arreglar/refactorizar
+- Archivo `chess/interfaz.py`
+- Archivo `chess/tablero.py`
+- Archivo `chess/piezas.py`
+- Archivo `chess/main.py`
+
+
+## [1.3.0] - 2024-08-22
+### Arreglar
+- Archivo `chess/piezas.py`
+- Archivo `chess/tablero.py`
+
 ## [1.2.1] - 2024-08-19
 ### Actualizar
 - Archivo `test/test_interfaz.py`